Modifico css y el theme provider
