+++ src/index.css
@@ -13,10 +13,26 @@
   -moz-osx-font-smoothing: grayscale;
 } */
 
+@import url('https://fonts.googleapis.com/css2?family=Outfit:wght@100..900&display=swap');
+
+:root {
+  --font-base: 'Outfit', sans-serif;
+}
+
+html {
+  font-family: var(--font-base);
+}
+
 #root {
   height: 100vh;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   background-color: #F1F6F5;
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
